test: MemberControllerTest customGiven 적용
